requirements.txt
Django
djangorestframework
python-openstackclient
python-swiftclient
requests
Jinja2==2.0
libvirt-python==1.3.0
boto
